Factor out generic `activate_background_task`
Factor out a function that, given the name of a background task,
activates it and runs it to completion (in a test context).
